mvt-loader: Add worker
