--- modules/mapbox-vector-tile/docs/README.md
@@ -1,20 +1,0 @@
-# @loaders.gl/mapbox-vector-tile
-
-The `@loaders.gl/mapbox-vector-tile` module handles the [Mapbox Vector Tile](https://github.com/mapbox/vector-tile-spec) format, a protobuf encoded format that defines geospatial geometries.
-
-## Installation
-
-```bash
-npm install @loaders.gl/mapbox-vector-tile
-npm install @loaders.gl/core
-```
-
-## Loaders and Writers
-
-| Loader                                                                                               |
-| ---------------------------------------------------------------------------------------------------- |
-| [`MapboxVectorTileLoader`](modules/mapbox-vector-tiles/docs/api-reference/mapbox-vector-tile-loader) |
-
-## Attribution
-
-The `MapboxVectorTileLoader` uses [`@mapbox/vector-tile`](https://github.com/mapbox/vector-tile-js) module under the BSD-3-Clause.
