--- modules/mapbox-vector-tile/src/index.js
@@ -1,1 +1,0 @@
-export {MapboxVectorTileLoader} from './mvt-loader';
