--- modules/mapbox-vector-tile/test/data/README.md
@@ -1,3 +1,0 @@
-# Data Attribution
-
-- gadm_usa_3-0-3.mvt from GADM which is free for non-comercial use
