--- modules/mapbox-vector-tile/test/index.js
@@ -1,3 +1,0 @@
-import './mvt/mvt-points.spec';
-import './mvt/mvt-lines.spec';
-import './mvt/mvt-polygons.spec';
