+++ modules/mvt/docs/README.md
@@ -1,0 +1,20 @@
+# @loaders.gl/mvt
+
+The `@loaders.gl/mvt` module handles the [Mapbox Vector Tile](https://github.com/mapbox/vector-tile-spec) format, a protobuf encoded format that defines geospatial geometries.
+
+## Installation
+
+```bash
+npm install @loaders.gl/mvt
+npm install @loaders.gl/core
+```
+
+## Loaders and Writers
+
+| Loader                                                    |
+| --------------------------------------------------------- |
+| [`MVTLoader`](modules/mvts/docs/api-reference/mvt-loader) |
+
+## Attribution
+
+The `MVTLoader` uses [`@mapbox/vector-tile`](https://github.com/mapbox/vector-tile-js) module under the BSD-3-Clause.
