+++ modules/mvt/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@loaders.gl/mapbox-vector-tile",
+  "name": "@loaders.gl/mvt",
   "description": "Loader for Mapbox Vector Tiles",
   "version": "2.1.0-alpha.1",
   "license": "MIT",
