+++ modules/mvt/src/index.js
@@ -1,0 +1,1 @@
+export {MVTLoader} from './mvt-loader';
