+++ modules/mvt/src/lib/parse-mvt.js
