+++ modules/mvt/src/workers/mvt-loader.worker.js
@@ -1,0 +1,4 @@
+import {MVTLoader} from './mvt-loader';
+import {createWorker} from '@loaders.gl/loader-utils';
+
+createWorker(MVTLoader);
