+++ modules/mvt/test/data/README.md
@@ -1,0 +1,3 @@
+# Data Attribution
+
+- gadm_usa_3-0-3.mvt from GADM which is free for non-commercial use
