+++ modules/mvt/test/data/lines_generated_5-16-11.mvt
