+++ modules/mvt/test/data/points_generated_2-0-1.mvt
