+++ modules/mvt/test/index.js
@@ -1,0 +1,1 @@
+import './mvt-loader.spec';
